patch to OpenMP parallel NBLIST, various other small bug fixes
git-svn-id: svn://dasher.wustl.edu/tinker@52 9dbd3ec4-0870-4ad8-bd6a-181c4f7deb04

+++ source/usolve.f
@@ -18,7 +18,7 @@ c
 c
       module usolve
       implicit none
-      real*8, allocatable :: mindex(:)
+      integer, allocatable :: mindex(:)
       real*8, allocatable :: minv(:)
       save
       end
